Don't show a delayed Tipsy if underlying element is removed
Currently if Tipsy has a `delayIn`, and the element is removed
from the DOM before the Tipsy displays, then the Tipsy will still
appear as an orphan at the top left of the page.  This change adds
an additional check to ensure that the element is in the DOM before
showing the Tipsy.